refactor(mercury): split unpacker.rs (864 lines) into directory module
Extract test module into sibling file:

- mod.rs: unpacker logic (520 lines)
- tests.rs: unit tests (344 lines)
